refactor(error): move error trait to be reused between crates
